Rename util
This file was renamed to `main`.

Task: SA3P-123
